fix: use new commit tag file for app version as well
