+++ server/routes/index.ts
@@ -1,5 +1,4 @@
 import { Router } from 'express';
-import path from 'path';
 import user from './user';
 import authRoutes from './auth';
 import { checkUser, isAuthenticated } from '../middleware/auth';
@@ -14,18 +13,7 @@ import tvRoutes from './tv';
 import mediaRoutes from './media';
 import personRoutes from './person';
 import collectionRoutes from './collection';
-import { getAppVersion } from '../utils/appVersion';
-import { existsSync } from 'fs';
-import logger from '../logger';
-
-const COMMIT_TAG_PATH = path.join(__dirname, '../../committag.json');
-let commitTag = 'local';
-
-if (existsSync(COMMIT_TAG_PATH)) {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  commitTag = require(COMMIT_TAG_PATH).commitTag;
-  logger.info(`Commit Tag: ${commitTag}`);
-}
+import { getAppVersion, getCommitTag } from '../utils/appVersion';
 
 const router = Router();
 
@@ -34,7 +22,7 @@ router.use(checkUser);
 router.get('/status', (req, res) => {
   return res.status(200).json({
     version: getAppVersion(),
-    commitTag,
+    commitTag: getCommitTag(),
   });
 });
